% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAWpaper-package.R
\docType{package}
\name{SAWpaper-package}
\alias{SAWpaper}
\alias{SAWpaper-package}
\title{SAWpaper: Self-Assessment Works Paper}
\description{
This package and corresponding GitHub repository are intended to
  enhance the reproducibility of a research paper. The principal investigators 
  for the study are Dr. Paula Winke (Associate Professor) and Dr. Susan Gass
 (University Distinguished Professor), both whom are at the Department of 
 Linguistics & Germanic, Slavic, Asian, and African Languages, Michigan State 
 University. The paper analyzes data on Spanish language learners who took a 
 Can-Do self-assessment test, along with the more authoritative OPIc language 
 proficiency test. We did both correlation analyses and continuation-ratio 
 models that examine the effect of course and OPIc speaking proficiency scores 
 on the passing rate for each level of the Can-Do statements self-assessment. 
 The objective was to validate the Can-Do test results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sjpierce/SAWpaper}
}

}
\author{
\strong{Maintainer}: Steven J. Pierce \email{pierces1@msu.edu} (\href{https://orcid.org/0000-0002-0679-3019}{ORCID}) [copyright holder]

}
\keyword{internal}
